# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 채용공고 Inbox API
  version: 0.3.0
  description: >
    Hand-maintained snapshot of the implemented MVP REST surface.
    Model: recruit-inbox-mvp-design-v1.1.md §7. Frontend types mirror this in
    frontend/lib/types.ts. springdoc auto-generation is deferred until a
    Spring Boot 4 compatible release ships. Production uses the Google/Kakao OAuth2
    session cookie; development may use X-Dev-User-Id. Mutating requests also
    require the X-XSRF-TOKEN header obtained from /api/v1/auth/csrf.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
  - devUser: []
...
paths:
...
  /api/v1/links:
    post:
      tags: [links]
      summary: Import a job URL (async). 202 new, 200 duplicate.
      parameters:
        - { name: Idempotency-Key, in: header, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties: { url: { type: string, maxLength: 4096 } }
      responses:
        "202": { description: queued, content: { application/json: { schema: { $ref: "#/components/schemas/LinkImportResponse" } } } }
        "200": { description: existing }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/links/{id}:
    get:
      tags: [links]
      summary: Get a saved link and its related applications and latest extraction
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: link + applications + latestRun }, "404": { $ref: "#/components/responses/Error" } }
  /api/v1/extractions/{id}:
    get:
      tags: [links]
      summary: Get an extraction run
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: run, content: { application/json: { schema: { $ref: "#/components/schemas/ExtractionRunResponse" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/applications:
    get:
      tags: [applications]
      summary: List applications
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ApplicationStatus" } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 30, maximum: 100 } }
      responses:
        "200": { description: page, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationPage" } } } }
    post:
      tags: [applications]
      summary: Create an application for a saved link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [linkId]
              properties:
                linkId: { type: string, format: uuid }
                positionKey: { type: string, maxLength: 100 }
                companyName: { type: string, maxLength: 200 }
                positionTitle: { type: string, maxLength: 300 }
                employmentType: { type: string, maxLength: 100 }
                experience: { type: string, maxLength: 100 }
                location: { type: string }
                notes: { type: string, maxLength: 10000 }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationResponse" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/applications/{id}:
    get:
      tags: [applications]
      summary: Get an application
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: application, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationResponse" } } } }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [applications]
      summary: Update an application
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties:
                expectedVersion: { type: integer }
                sourceUrl: { type: string, format: uri, maxLength: 4096 }
                companyName: { type: string, maxLength: 200 }
                positionTitle: { type: string, maxLength: 300 }
                employmentType: { type: string, maxLength: 100 }
                experience: { type: string, maxLength: 100 }
                location: { type: string }
                status: { $ref: "#/components/schemas/ApplicationStatus" }
                notes: { type: string, maxLength: 10000 }
                archived: { type: boolean }
      responses:
        "200": { description: updated, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationResponse" } } } }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [applications]
      summary: Delete an application
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: If-Match, in: header, required: true, schema: { type: string } }
      responses: { "204": { description: deleted }, "412": { $ref: "#/components/responses/Error" }, "428": { $ref: "#/components/responses/Error" } }
  /api/v1/applications/manual:
    post:
      tags: [applications]
      summary: Create a manually entered application with an optional source URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyName, positionTitle]
              properties:
                companyName: { type: string, maxLength: 200 }
                positionTitle: { type: string, maxLength: 300 }
                sourceUrl: { type: string, format: uri, maxLength: 4096 }
                notes: { type: string, maxLength: 10000 }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/ApplicationResponse" } } } }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/applications/{id}/confirm:
    post:
      tags: [applications]
      summary: Confirm extracted application fields and selected event proposals
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties:
                expectedVersion: { type: integer }
                runId: { type: string, format: uuid }
                companyName: { type: string }
                positionTitle: { type: string }
                placeholderCandidates:
                  type: array
                  items:
                    type: object
                    required: [type]
                    properties:
                      type: { $ref: "#/components/schemas/EventType" }
                      customLabel: { type: string }
                      sourceCandidateId: { type: string }
      responses: { "200": { description: application + createdEvents }, "409": { $ref: "#/components/responses/Error" } }
  /api/v1/applications/{id}/essay-questions:
    get:
      tags: [essays]
      summary: List cover-letter questions and current drafts for an application
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: Questions in display order
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/EssayQuestionResponse" } }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [essays]
      summary: Add a cover-letter question
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EssayQuestionCreate" }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/EssayQuestionResponse" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/essay-questions/{id}:
    patch:
      tags: [essays]
      summary: Update a question and its current draft
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EssayQuestionUpdate" }
      responses:
        "200": { description: updated, content: { application/json: { schema: { $ref: "#/components/schemas/EssayQuestionResponse" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [essays]
      summary: Delete a question and all of its saved revisions
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: If-Match, in: header, required: true, schema: { type: string } }
      responses:
        "204": { description: deleted }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/essay-questions/{id}/revisions:
    get:
      tags: [essays]
      summary: List saved answer revisions, newest first
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: revisions
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/EssayRevisionResponse" } }
    post:
      tags: [essays]
      summary: Snapshot the currently saved question and answer
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties:
                expectedVersion: { type: integer, format: int64 }
                label: { type: string, maxLength: 100 }
      responses:
        "201": { description: revision created, content: { application/json: { schema: { $ref: "#/components/schemas/EssayRevisionResponse" } } } }
  /api/v1/essay-questions/{id}/revisions/{revisionId}/restore:
    post:
      tags: [essays]
      summary: Restore a saved answer as the current draft
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: revisionId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              properties: { expectedVersion: { type: integer, format: int64 } }
      responses:
        "200": { description: restored current draft, content: { application/json: { schema: { $ref: "#/components/schemas/EssayQuestionResponse" } } } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Spring Session cookie established by Google or Kakao OAuth2 login in production.
    devUser:
      type: apiKey
      in: header
      name: X-Dev-User-Id
      description: Development only; the UUID must identify an existing local user.
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
                  retryable: { type: boolean }
                  requestId: { type: string }
                  fields: { type: object, additionalProperties: { type: string } }
  schemas:
...
    ApplicationStatus:
      type: string
      enum: [INTERESTED, PLANNED, APPLIED, IN_PROGRESS, ACCEPTED, REJECTED, WITHDRAWN]
    EventType:
      type: string
      enum: [DOCUMENT_DEADLINE, NCS, CODING_TEST, AI_ASSESSMENT, INTERVIEW_1, INTERVIEW_2, FINAL_INTERVIEW, RESULT_ANNOUNCEMENT, ORIENTATION, CUSTOM]
...
    LinkImportResponse:
      type: object
      properties:
        linkId: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        extractionRunId: { type: [string, "null"], format: uuid }
        status: { type: string }
        duplicate: { type: boolean }
    ExtractionRunResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [QUEUED, RUNNING, SUCCEEDED, NEEDS_INPUT, FAILED, CANCELLED] }
        progressStage: { type: [string, "null"] }
        attemptCount: { type: integer }
        result: { type: object }
        warnings: { type: array, items: {} }
        errorCode: { type: [string, "null"] }
    ApplicationResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        linkId: { type: string, format: uuid }
        sourceUrl: { type: [string, "null"], format: uri }
        positionKey: { type: string }
        companyName: { type: [string, "null"] }
        positionTitle: { type: [string, "null"] }
        employmentType: { type: [string, "null"] }
        experience: { type: [string, "null"] }
        location: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        reviewStatus: { type: string, enum: [PENDING, CONFIRMED, NOT_REQUIRED] }
        appliedAt: { type: [string, "null"], format: date-time }
        notes: { type: string }
        archivedAt: { type: [string, "null"], format: date-time }
        version: { type: integer }
        fieldMeta: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ApplicationPage:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ApplicationResponse" } }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
    EssayLimitType:
      type: string
      enum: [NONE, CHARACTERS_WITH_SPACES, CHARACTERS_WITHOUT_SPACES, UTF8_BYTES, KOREAN_2_BYTES]
    EssayQuestionCreate:
      type: object
      required: [questionText]
      properties:
        questionText: { type: string, minLength: 1, maxLength: 5000 }
        limitType: { $ref: "#/components/schemas/EssayLimitType" }
        limitValue: { type: integer, minimum: 1, maximum: 1000000 }
    EssayQuestionUpdate:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64 }
        questionText: { type: string, minLength: 1, maxLength: 5000 }
        limitType: { $ref: "#/components/schemas/EssayLimitType" }
        limitValue: { type: integer, minimum: 1, maximum: 1000000 }
        answerText: { type: string, maxLength: 100000 }
        status: { type: string, enum: [DRAFT, COMPLETED] }
    EssayQuestionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        sortOrder: { type: integer }
        questionText: { type: string }
        limitType: { $ref: "#/components/schemas/EssayLimitType" }
        limitValue: { type: [integer, "null"] }
        answerText: { type: string }
        status: { type: string, enum: [DRAFT, COMPLETED] }
        characterCount: { type: integer }
        characterCountWithoutSpaces: { type: integer }
        utf8ByteCount: { type: integer }
        korean2ByteCount: { type: integer }
        revisionCount: { type: integer, format: int64 }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    EssayRevisionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        questionId: { type: string, format: uuid }
        revisionNo: { type: integer }
        label: { type: [string, "null"] }
        questionText: { type: string }
        limitType: { $ref: "#/components/schemas/EssayLimitType" }
        limitValue: { type: [integer, "null"] }
        answerText: { type: string }
        characterCount: { type: integer }
        characterCountWithoutSpaces: { type: integer }
        utf8ByteCount: { type: integer }
        korean2ByteCount: { type: integer }
        createdAt: { type: string, format: date-time }
...